backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "audiorailrider-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "myrqyry",email = "matrmx@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "google-genai==1.28.0",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "pydantic-settings (>=2.2.1,<3.0.0)",
    "slowapi (>=0.1.9,<1.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-mock = "^3.0.0"
